requirements/requirements.txt
typer>=0.12.5 
rich-cli>=1.8.0